+++ saxon/src/test/java/org/teic/teixptr/implementation/TestSetup.java
@@ -24,9 +24,9 @@ public class TestSetup {
 
     public static String mtXml = Paths.get(realworldSamples.resolve("MT.tei.xml")).toFile().toString();
 
-    public static String satsxp01, satsr01, satsl01, satssi01, satsrn01;
+    public static String satsxp01, satsr01, satsl01, satssi01, satsrn01, satssr01, satssr02, satssr03;
 
-    public static String sysidref01, sysidref02, sysxp02, sysrn04, sysrn05, sysrn06, sysrn07, sysrn08, sysrn09;
+    public static String sysidref01, sysidref02, sysxp02, sysxp03, sysxp04, sysrn04, sysrn05, sysrn06, sysrn07, sysrn08, sysrn09, syssi02;
 
     public static String ijobidref01, ijobrn01;
 
@@ -52,6 +52,9 @@ public class TestSetup {
 	satsl01 = readSatsSample("satsl01.txt");
 	satssi01 = readSatsSample("satssi01.txt");
 	satsrn01 = readSatsSample("satsrn01.txt");
+	satssr01 = readSatsSample("satssr01.txt");
+	satssr02 = readSatsSample("satssr02.txt");
+	satssr03 = readSatsSample("satssr03.txt");
 
 
 	sysidref01 = readSystematicSample("idref01.txt");
@@ -63,6 +66,9 @@ public class TestSetup {
 	sysrn08 = readSystematicSample("rn08.txt");
 	sysrn09 = readSystematicSample("rn09.txt");
 	sysxp02 = readSystematicSample("xp02.txt");
+	sysxp03 = readSystematicSample("xp03.txt");
+	sysxp04 = readSystematicSample("xp04.txt");
+	syssi02 = readSystematicSample("si02.txt");
 
 	ijobidref01 = readRealWorldSample("ijobidref01.txt");
 	ijobrn01 = readRealWorldSample("ijobrn01.txt");